fix ds.box on windows

+++ yt/data_objects/static_output.py
@@ -964,9 +964,11 @@ class Dataset(object):
         # list or other non-array iterable before calculating
         # the center
         if not isinstance(left_edge, np.ndarray):
-            left_edge = np.array(left_edge, dtype='float64')
+            left_edge = np.array(left_edge)
         if not isinstance(right_edge, np.ndarray):
-            right_edge = np.array(right_edge, dtype='float64')
+            right_edge = np.array(right_edge)
+        left_edge = left_edge.astype('float64')
+        right_edge = right_edge.astype('float64')
         c = (left_edge + right_edge)/2.0
         return self.region(c, left_edge, right_edge, **kwargs)
